+++ src/components/atoms/buttons/ButtonCategory/index.spec.tsx
@@ -3,7 +3,7 @@ import { mockCategories } from "mocks/data/categories";
 
 import { urlTable } from "@/utils/paths/url";
 
-import ButtonCategory from ".";
+import { ButtonCategory } from ".";
 
 describe("components/atoms/buttons/ButtonCategory", () => {
   const categoryRice = mockCategories.rice;
