+++ src/components/atoms/buttons/index.ts
@@ -1,0 +1,3 @@
+export { ButtonCategory } from "./ButtonCategory";
+export { ButtonTagColored, ButtonTagPlain } from "./ButtonTag";
+export { LoadMoreButton } from "./LoadMoreButton";
