+++ src/components/molecules/TagList/Colored/index.tsx
@@ -1,4 +1,4 @@
-import { ButtonTagColored } from "@/components/atoms/buttons/ButtonTag";
+import { ButtonTagColored } from "@/components/atoms/buttons";
 import type { TTag } from "@/types";
 
 type Props = {
