+++ src/components/pages/articles/ArticleDetail/index.tsx
@@ -2,7 +2,7 @@ import { MDXRemote } from "next-mdx-remote";
 import type { MDXRemoteSerializeResult } from "next-mdx-remote/dist/types";
 import { FaPen, FaRegCalendar } from "react-icons/fa";
 
-import ButtonCategory from "@/components/atoms/buttons/ButtonCategory";
+import { ButtonCategory } from "@/components/atoms/buttons";
 import MDXCustomComponents from "@/components/atoms/mdx";
 import TextDate from "@/components/atoms/texts/TextDate";
 import Thumbnail from "@/components/atoms/Thumbnail";
